Moved TestSettingsProvider to the new commonTest package
