+++ collect_app/src/androidTest/java/org/odk/collect/android/support/TestUtils.java
@@ -9,11 +9,11 @@ import androidx.test.espresso.util.HumanReadables;
 import androidx.test.espresso.util.TreeIterables;
 
 import org.hamcrest.Matcher;
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.database.DatabaseInstancesRepository;
 import org.odk.collect.android.preferences.source.Settings;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import java.io.Closeable;
 import java.io.File;
