+++ collect_app/src/test/java/org/odk/collect/android/application/initialization/migration/KeyRenamerTest.java
@@ -3,8 +3,8 @@ package org.odk.collect.android.application.initialization.migration;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.source.Settings;
-import org.odk.collect.utilities.TestSettingsProvider;
 import org.robolectric.RobolectricTestRunner;
 
 import static org.odk.collect.android.application.initialization.migration.MigrationUtils.renameKey;
