+++ collect_app/src/test/java/org/odk/collect/android/application/initialization/migration/SharedPreferenceUtils.java
@@ -1,7 +1,7 @@
 package org.odk.collect.android.application.initialization.migration;
 
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.source.Settings;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import static org.junit.Assert.assertEquals;
