+++ collect_app/src/test/java/org/odk/collect/android/configure/SettingsImporterTest.java
@@ -7,9 +7,9 @@ import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.application.initialization.SettingsPreferenceMigrator;
 import org.odk.collect.android.preferences.source.Settings;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import java.util.ArrayList;
 import java.util.HashMap;
