+++ collect_app/src/test/java/org/odk/collect/android/configure/qr/QRCodeViewModelTest.java
@@ -6,9 +6,9 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.R;
 
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.source.Settings;
 import org.odk.collect.testshared.FakeScheduler;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import static java.util.Arrays.asList;
 import static org.hamcrest.MatcherAssert.assertThat;
