+++ collect_app/src/test/java/org/odk/collect/android/preferences/screens/MainMenuAccessPreferencesTest.java
@@ -10,12 +10,12 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.FormUpdateMode;
 import org.odk.collect.android.preferences.keys.GeneralKeys;
 import org.odk.collect.android.preferences.source.Settings;
 import org.odk.collect.android.preferences.Protocol;
 import org.odk.collect.android.preferences.screens.AdminPreferencesFragment.MainMenuAccessPreferences;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
