+++ collect_app/src/test/java/org/odk/collect/android/tasks/InstanceSyncTaskTest.java
@@ -4,11 +4,11 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.database.DatabaseInstancesRepository;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.odk.collect.android.support.InstanceUtils;
-import org.odk.collect.utilities.TestSettingsProvider;
 
 import java.io.File;
