+++ collect_app/src/test/java/org/odk/collect/android/widgets/base/WidgetTest.java
@@ -11,8 +11,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
+import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.source.SettingsProvider;
-import org.odk.collect.utilities.TestSettingsProvider;
 import org.robolectric.RobolectricTestRunner;
 
 import javax.annotation.OverridingMethodsMustInvokeSuper;
